Add cache module for HistoricalDataCache
